Refactor labels series layer but broken
